HydroSure_backend/requirements.txt
langchain
langgraph
langchain-google-genai
numpy
Pillow